fixed 0c64323 from https://gitee.com/onexiaomin/multimedia_medialibrary_standard/pulls/1004 medialibrary codex clean

+++ frameworks/innerkitsimpl/test/unittest/get_self_permissions/src/get_self_permissions.cpp
@@ -27,7 +27,7 @@ void PermissionUtilsUnitTest::SetAccessTokenPermission(const std::string &proces
     const std::vector<std::string> &permission, uint64_t &tokenId)
 {
     auto perms = std::make_unique<const char *[]>(permission.size());
-    for (int i = 0; i < permission.size(); i++) {
+    for (size_t i = 0; i < permission.size(); i++) {
         perms[i] = permission[i].c_str();
     }
